Fix mistakes and add more work in backend

+++ backend/src/main/java/com/xiojuandawt/blood4life/repositories/AppointmentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.xiojuandawt.blood4life.entities.Appointment;
 
-public interface AppointmentRepository extends JpaRepository<Appointment, Long> {
-    Long countByCampaignId(Integer campaignId);
+public interface AppointmentRepository extends JpaRepository<Appointment, Integer> {
+    Integer countByCampaignId(Integer campaignId);
 }
